@@ -3,14 +3,16 @@ package SuperChatDeluxe.model;
 import java.io.Serializable;
 import java.time.LocalDateTime;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+
+
 
 @Entity
 @Table(name = "message")
@@ -36,8 +38,7 @@ public class Message implements Serializable {
 	private User username;
 	
 	@ManyToOne
-	@JoinColumn(name = "sent_to", referencedColumnName = "username")
-	@Column(nullable = true)
+	@JoinColumn(name = "sent_to", referencedColumnName = "username", nullable = true)
 	private User sentTo;
 	
 	public Message() {
